Refactoring package to distribute, initial setup.py using setuptools
We have some problem when removing the package... pip uninstall fails
in that point.
By other hand, the entry point is working fine.

I need only setup the initial .myanimelist.init. If possible, would
be good avoid savind the credentials in plain text. But the API of
MyAnimeList is very fucked...