Create new aliases for inc/dec +1/-1, create utils.py and login.py
Score 0 means not scored yet, for that reason, I tag on the output as
'undefined'. In MAL scores doesn't exists the score 0 (only in the API).

The login.py is used for the general function to handle the credentials.
If the file was not found in the default directory, the script suppose
that file doesn't exists so create a new, asking the user/password and
later saving in the default path.