PyPI prepare: fixes about setup.py and add full-featured makefile script
The setup.py now requires pypandoc for execute properly and send the
long_description as reStructuredText on PyPI. If the users don't have,
continues the build, but show a warning about this.

I refactored the general classifiers to a more semantic way too.

The makefile now can handle that entries:

* make install # install default
* make develop # develop mode with symlinks
* make test-register # register in testpypi server
* make test-deploy # deply to the testpypi server
* make register # register in pypi server
* make deploy # deploy release in pypi server
* make check # check the package about errors (metadata, readme)
* make clean # clean build trashs