Add a hacky @spinnered decorator to print animation while is running
Well, is a lot of hacks, someones ugly, others really funny. I need
revisit this later, but for now are working. I will provide a cool
gif to put in the README.md later demonstrating that functionality.

I have some trouble with handle the threads, because another
function @spinnered can be called and we not must call another animation
beyond that, when occurs some exception, the @checked_connection must
finish the last thread. This is really a problem. In the start, I was
with a lot of global variables, but for now I need only kill one later:
'last_thread'.