Change spinner animation & Regex Exception Handling
Now the spinner animation have two differents animations and the
content of message is conditional for the signal, normally the
name of function whose is running.

Besides, I change the name of @spinnered decorator to @animated.
I think is more clear what is does.

The regex exception handling is made inside of MyAnimeList.find().
It decorated with @checked_regex decorator, whose try/catch errors.
I implement the @checked_regex in this commit.

Maybe I need create a decorator generator for checked type later