Look for login command flag before trying to login.get_credentials
This avoid the weird behavior of a new user, without config file and
login credentials, try the first time use the application and call
`mal login`. Without this conditional the login.get_credentials()
will be called first, and as the user don't have it, will be ask to
create.

However, the user called too the `login` command, which in the final
line of main will be called on args.func(mal, args) and that will call
again the login.create_credentials!

I added to the add_subparsers the kwarg dest='command' to return
on args namespace on the end which command was signaled for the user.
That way is more simple to detect the command used. Another
alternative was using args.func.__name__, but this is ugly.