[cli] translate anime_regex to anime_title on search command
This would avoid confusion since global search command doesn't use
regex at all! This just happened because the old search command
was renamed to filter, which use regex to match on the local
userlist.

Since we search is done by the REST API of mal, and I don't have
intention to do horrible web scrap stuff on something like mal,
this is the thing.

Being explicit it's important, sorry for the users that though
that would be possible search GLOBALLY for animes on MyAnimeList.
This will not gonna happen.