Readme.md overhaul v2
Small edit. This commit message is a copy of the previous one.

I tried to keep the lines around 140 characters at the most, excluding
things that'll be collapsed by the parser, like urls.

I also decided to add a `$` to the beginning of every shell command.
This makes it possible to add command output to an example without it
being confusing. It's clear which lines are commands and which lines
are output.

This is how I plan to format the rest of the docs, where applicable.
If you have any notes, now is the time.

Also note: In the spirit of good practice, this commit message wraps
at 72 cols.